Fix worng MVP architecture.
